Move object_new and related functions to objects
